2.0.0.beta04: massive update. adaptive bisection is now ok with mixed trials
all trials have the new stim_value properties that store the value given to the subject

massive refactoring and optimization of TestFragment and TestBasic

---TestBasic.testEvent
now emit a Triple(int, any, List<String>)
added several events that are triggered in TestFragment
EVENT_TRIAL_STARTED
EVENT_NAVIGATE_BACK

added:
- AdaptiveTrialsManager
- BisAdaptiveTrialsManager
- MixedAdaptiveTrialsManager

BIS:

python code refactored. before/after midline is managed in kotlin

added magnitude properties in TrialBIS and added TrialBISADA
the idea is to deal with a magnitude values, set statically or adaptively calculated at runtime, and then apply it before or after midline latency

+++ psysuitecore/src/main/java/iit/uvip/psysuite/core/tests/temporalbinding/TrialBindingsUnBalanced.kt
@@ -25,7 +25,7 @@ class TrialBindingsUnBalanced(id:Int=-1, type:Int=0, override var stim_value:Lon
         return "${super.debugInfo()}, delay=$stim_value"
     }
 
-    override fun updateTrial(newvalue:Float){
-        stim_value = newvalue.toLong()
+    override fun updateTrial(newvalue:Long){
+        stim_value = newvalue
     }
 }
